Update README.md
I'm looking through some PIPs, but it's a bit burdensome to know what's where. 

I added a table to the README- is this helpful for quickly browsing the repo? It's a bit of a manual process so I only wrote out the first 10 PIPs.  If you like the change and merge, I'll add the rest. 

Not sure if the table will make the README too long or bury PM info at the bottom? I considered writing a new doc to the repo with a table added, but that didn't seem like the best UX choice.